python/requirements.txt
absl-py==0.4.1
astor==0.7.1
click==6.7
config==0.4.0
Flask==1.0.2
Flask-MySQLdb==0.2.0
Flask-SocketIO==3.0.2
Flask-Sockets==0.2.1
future==0.16.0
gast==0.2.0
gevent==1.3.6
gevent-websocket==0.10.1
greenlet==0.4.15
grpcio==1.14.2
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
msgpack-python==0.5.6
mysqlclient==1.3.13
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.3.0
protobuf==3.6.1
python-engineio==2.2.0
python-socketio==2.0.0
PyYAML==3.13
pyzmq==17.1.2
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
Werkzeug==0.14.1
zerorpc==0.6.1
